fix: missing advance metrics recalc on L2 transfer tools
