+++ src/mappings/helpers/helpers.ts
@@ -808,6 +808,23 @@ export function updateDelegationExchangeRate(indexer: Indexer): Indexer {
   return indexer as Indexer
 }
 
+// TODO - this is broken if we change the delegatio ratio
+// Need to remove, or find a fix
+export function calculateCapacities(indexer: Indexer): Indexer {
+  let graphNetwork = GraphNetwork.load('1')!
+  let tokensDelegatedMax = indexer.stakedTokens.times(BigInt.fromI32(graphNetwork.delegationRatio))
+
+  // Eligible to add to the capacity
+  indexer.delegatedCapacity =
+    indexer.delegatedTokens < tokensDelegatedMax ? indexer.delegatedTokens : tokensDelegatedMax
+
+  indexer.tokenCapacity = indexer.stakedTokens.plus(indexer.delegatedCapacity)
+  indexer.availableStake = indexer.tokenCapacity
+    .minus(indexer.allocatedTokens)
+    .minus(indexer.lockedTokens)
+  return indexer
+}
+
 export function calculatePricePerShare(deployment: SubgraphDeployment): BigDecimal {
   // TODO check why there's a deviation from the values of the bancor formula
   // Ideally this would be a 1 to 1 recreation of the share sell formula, but due to
